feat/tratamento-de-excecoes

+++ Desafio2_MicroserviceB/src/main/java/com/compass/Desafio2_MicroserviceB/service/PostService.java
@@ -1,17 +1,20 @@
 package com.compass.Desafio2_MicroserviceB.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.stereotype.Service;
-
 import com.compass.Desafio2_MicroserviceB.client.JsonPlaceholderClient;
 import com.compass.Desafio2_MicroserviceB.dto.PostDTO;
+import com.compass.Desafio2_MicroserviceB.exceptions.CamposInvalidosException;
+import com.compass.Desafio2_MicroserviceB.exceptions.EntityNotFoundException;
 import com.compass.Desafio2_MicroserviceB.mapper.PostMapper;
 import com.compass.Desafio2_MicroserviceB.model.Post;
 import com.compass.Desafio2_MicroserviceB.repository.PostRepository;
-
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -26,6 +29,7 @@ public class PostService {
         postDTOList.forEach(this::savePostFromDTO);
     }
 
+    @Transactional(readOnly = true)
     public List<PostDTO> getAllPosts() {
         List<Post> posts = postRepository.findAll();
         return posts.stream()
@@ -33,25 +37,37 @@ public class PostService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional(readOnly = true)
     public PostDTO getPostById(Long id) {
         return postRepository.findById(id)
                 .map(postMapper::convertToDTO)
-                .orElseThrow(() -> new RuntimeException("Post não encontrado"));
+                .orElseThrow(() -> new EntityNotFoundException("Post não encontrado"));
     }
 
-    public PostDTO createPost(PostDTO postDTO) {
-        Post post = postMapper.convertToEntity(postDTO);
-        return postMapper.convertToDTO(postRepository.save(post));
+    public PostDTO createPost(@Valid PostDTO postDTO) {
+        try {
+            Post post = postMapper.convertToEntity(postDTO);
+            return postMapper.convertToDTO(postRepository.save(post));
+        } catch (ConstraintViolationException e) {
+            throw new CamposInvalidosException("Os campos do post não podem estar vazios.");
+        }
     }
 
-    public PostDTO updatePost(Long id, PostDTO postDTO) {
+    public PostDTO updatePost(Long id, @Valid PostDTO postDTO) {
         Post post = postRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Post não encontrado"));
-        updatePostEntity(post, postDTO);
-        return postMapper.convertToDTO(postRepository.save(post));
+                .orElseThrow(() -> new EntityNotFoundException("Post não encontrado"));
+        try {
+            updatePostEntity(post, postDTO);
+            return postMapper.convertToDTO(postRepository.save(post));
+        } catch (Exception e) {
+            throw new CamposInvalidosException("Os campos do post não podem estar vazios.");
+        }
     }
 
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id))
+            throw new EntityNotFoundException("Post não encontrado");
+
         postRepository.deleteById(id);
     }
